quick refactors
