+++ library/src/main/java/com/mux/stats/sdk/muxstats/PlayerUtils.kt
@@ -1,28 +1,29 @@
-package com.mux.stats.sdk.muxstats.internal
+package com.mux.stats.sdk.muxstats
 
 import android.util.Log
 import androidx.media3.common.Player
 import androidx.media3.common.Tracks
 import com.mux.android.util.oneOf
-import com.mux.stats.sdk.muxstats.MuxPlayerState
-import com.mux.stats.sdk.muxstats.MuxStateCollector
-import java.lang.ref.WeakReference
 
 internal const val PLAYER_STATE_POLL_MS = 150L
 
-internal fun watchPlayerPos(player: Player, collector: MuxStateCollector) {
-  collector.playerWatcher = MuxStateCollector.PlayerWatcher(
+/**
+ * Asynchronously watch player playback position, collecting periodic updates out-of-band from the
+ * normal callback flow.
+ */
+fun MuxStateCollector.watchPlayerPos(player: Player) {
+  playerWatcher = MuxStateCollector.PlayerWatcher(
     PLAYER_STATE_POLL_MS,
-    collector,
+    this,
     player
   ) { it, _ -> it.currentPosition }
-  collector.playerWatcher?.start()
+  playerWatcher?.start()
 }
 
 /**
  * Returns true if any media track in the given [Tracks] object had a video MIME type
  */
-internal fun Tracks.hasAtLeastOneVideoTrack(): Boolean {
+fun Tracks.hasAtLeastOneVideoTrack(): Boolean {
   return groups.map { it.mediaTrackGroup }
     .filter { trackGroup -> trackGroup.length > 0 }
     .map { trackGroup -> trackGroup.getFormat(0) }
@@ -34,7 +35,7 @@ internal fun Tracks.hasAtLeastOneVideoTrack(): Boolean {
  * Handles an ExoPlayer position discontinuity
  */
 @JvmSynthetic // Hides from java
-internal fun MuxStateCollector.handlePositionDiscontinuity(reason: Int) {
+fun MuxStateCollector.handlePositionDiscontinuity(reason: Int) {
   when (reason) {
     Player.DISCONTINUITY_REASON_SEEK -> {
       // If they seek while paused, this is how we know the seek is complete
@@ -56,7 +57,7 @@ internal fun MuxStateCollector.handlePositionDiscontinuity(reason: Int) {
  * Handles a change of basic ExoPlayer state
  */
 @JvmSynthetic // Hidden from Java callers, since the only ones are external
-internal fun MuxStateCollector.handleExoPlaybackState(
+fun MuxStateCollector.handleExoPlaybackState(
   playbackState: Int, // the @IntDef for player state omitted. Unavailable on all exo versions
   playWhenReady: Boolean
 ) {
